requirements.txt
aiobotocore==2.5.4
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.2
anyio==3.7.1
appdirs==1.4.4
apprise==1.4.5
asgi-lifespan==2.1.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.20
botocore==1.31.17
bs4==0.0.1
build==0.10.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.4.1
cryptography==41.0.3
cssselect==1.2.0
dataclasses-json==0.5.14
dateparser==1.1.8
Django==4.2.4
dnspython==2.4.2
docker==6.1.3
fake-useragent==1.2.1
fastapi==0.101.0
frozenlist==1.4.0
fsspec==2023.6.0
google-auth==2.22.0
greenlet==2.0.2
griffe==0.32.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kubernetes==27.2.0
langchain==0.0.273
langsmith==0.0.26
loguru==0.7.0
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-boto3-s3==1.28.19
mypy-boto3-secretsmanager==1.28.16
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.9
orjson==3.9.2
packaging==23.1
parse==1.19.1
pathspec==0.11.2
pendulum==2.1.2
pinecone-client==2.2.2
postgres==4.0
prefect==2.11.3
prefect-aws==0.3.6
psycopg2==2.9.7
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pyee==8.2.2
Pygments==2.15.1
PyJWT==2.8.0
pyppeteer==1.0.2
pyproject_hooks==1.0.0
pyquery==2.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
pywin32==306
PyYAML==6.0.1
readchar==4.0.5
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3fs==2023.6.0
s3transfer==0.6.1
singlestoredb==0.8.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparams==5.1.0
sqlparse==0.4.4
starlette==0.27.0
tenacity==8.2.3
text-unidecode==1.3
tiktoken==0.4.0
toml==0.10.2
tqdm==4.65.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.23.2
w3lib==2.1.2
websocket-client==1.6.1
websockets==10.4
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
